// angular/src/app/shared/model/types.ts
export enum DataTypes {

    Vendor = "vendor",
    Client = "client",
    Community = "community",
    OrganisationTranslation = "organisationtranslation",
    SuccessStory = "successstory",
    SuccessStoryTranslation = "successstorytranslation",
    Product = "product",
    ProductTranslation = "producttranslation",
    Category = "category",
    Address = "address",
    VendorServices = "vendorservices"

}